# {{cookiecutter.project_slug}}/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API documentation for {{ cookiecutter.project_name }}.",
    "title": "{{ cookiecutter.project_name }} API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "email": "support@{{ cookiecutter.project_name }}.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
    "/account/get-account": {
      "get": {
        "description": "Retrieve user account information",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Get user account",
        "parameters": [
          {
            "description": "Get user account details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.GetUserReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account retrieved successfully",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/internal_users.GetUserRes"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Initiate forgot password process",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Forgot password",
        "parameters": [
          {
            "description": "Forgot password details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If your email is registered, you will receive a verification code",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate user and return access token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Login a user",
        "parameters": [
          {
            "description": "Login details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/internal_users.LoginResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "401": {
            "description": "Invalid email or password",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "403": {
            "description": "Email not verified",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "Create a new user account",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Register a new user",
        "parameters": [
          {
            "description": "Registration details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Account created successfully",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/internal_users.UserResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "409": {
            "description": "Email or phone already exists",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Reset user password with verification code",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Reset password",
        "parameters": [
          {
            "description": "Reset password details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.ResetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "404": {
            "description": "Invalid or expired verification code",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "description": "Verify user email with verification code",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Verify user email",
        "parameters": [
          {
            "description": "Verification details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_users.VerifyEmailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          },
          "404": {
            "description": "Invalid or expired verification code",
            "schema": {
              "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Meta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "total_count": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        }
      }
    },
    "github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {
          "type": "string"
        },
        "error_code": {
          "type": "string"
        },
        "errors": {},
        "message": {
          "type": "string"
        },
        "meta": {
          "$ref": "#/definitions/github_com_{{ cookiecutter.github_username }}_{{ cookiecutter.project_name }}_pkg_response.Meta"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
    "internal_users.ForgotPasswordRequest": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "internal_users.GetUserReq": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "internal_users.GetUserRes": {
      "type": "object",
      "properties": {
        "age": {
          "type": "integer"
        },
        "avatar": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "internal_users.LoginRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "internal_users.LoginResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "access_token_expires_at": {
          "type": "string"
        },
        "refresh_token": {
          "type": "string"
        },
        "refresh_token_expires_at": {
          "type": "string"
        },
        "session_id": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/internal_users.UserResponse"
        }
      }
    },
    "internal_users.RegisterRequest": {
      "type": "object",
      "required": [
        "account_type",
        "email",
        "first_name",
        "last_name",
        "password",
        "phone_number"
      ],
      "properties": {
        "account_type": {
          "type": "string",
          "enum": ["user", "merchant"]
        },
        "email": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "minLength": 8
        },
        "phone_number": {
          "type": "string"
        }
      }
    },
    "internal_users.ResetPasswordRequest": {
      "type": "object",
      "required": ["code", "email", "new_password"],
      "properties": {
        "code": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "new_password": {
          "type": "string",
          "minLength": 8
        }
      }
    },
    "internal_users.UserResponse": {
      "type": "object",
      "properties": {
        "account_type": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "email_verified": {
          "type": "boolean"
        },
        "first_name": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        }
      }
    },
    "internal_users.VerifyEmailRequest": {
      "type": "object",
      "required": ["code", "email"],
      "properties": {
        "code": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "Use your API key to authenticate requests.",
      "type": "apiKey",
      "name": "api-access-key",
      "in": "header"
    },
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}